refactor(stewardship): split tests.rs (#1266)
tests.rs 554 -> 306 LOC; new tests_extra.rs (~377 LOC).
