+++ src/stewardship/tests.rs
@@ -303,252 +303,3 @@ fn process_run_files_new_when_no_match() {
     assert_eq!(item.source, "stewardship:rysweet/Simard#42");
     assert!(item.description.contains("42") || item.description.contains(&sig));
 }
-
-#[test]
-fn process_run_matches_existing_when_signature_present() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let run = sample_run();
-    let sig = failure_signature(&run.failure_kind, &run.error_text);
-
-    gh.seed_search(
-        "rysweet/Simard",
-        &sig,
-        Ok(vec![GhIssue {
-            number: 11,
-            url: "https://github.com/rysweet/Simard/issues/11".into(),
-            title: "[stewardship] previously filed".into(),
-            body: format!("stewardship-signature: {sig}\n## Error\nold"),
-        }]),
-    );
-
-    let outcome = process_orchestrator_run(&run, &gh, &mut board).unwrap();
-    match outcome {
-        StewardshipOutcome::MatchedExisting {
-            issue_number, repo, ..
-        } => {
-            assert_eq!(issue_number, 11);
-            assert_eq!(repo, "rysweet/Simard");
-        }
-        other => panic!("expected MatchedExisting, got {other:?}"),
-    }
-
-    assert_eq!(gh.search_call_count(), 1);
-    assert_eq!(
-        gh.create_call_count(),
-        0,
-        "must NOT create when match exists"
-    );
-    assert_eq!(board.backlog.len(), 1);
-    assert_eq!(board.backlog[0].id, "stewardship-rysweet_Simard-11");
-}
-
-#[test]
-fn process_run_idempotent_on_second_invocation() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let run = sample_run();
-    let sig = failure_signature(&run.failure_kind, &run.error_text);
-
-    // First call: empty search → file new (#42).
-    gh.seed_search("rysweet/Simard", &sig, Ok(vec![]));
-    gh.seed_create(
-        "rysweet/Simard",
-        Ok(GhIssue {
-            number: 42,
-            url: "https://github.com/rysweet/Simard/issues/42".into(),
-            title: "t".into(),
-            body: format!("stewardship-signature: {sig}"),
-        }),
-    );
-    let first = process_orchestrator_run(&run, &gh, &mut board).unwrap();
-    assert!(matches!(
-        first,
-        StewardshipOutcome::FiledNew {
-            issue_number: 42,
-            ..
-        }
-    ));
-
-    // Second call: search now returns the issue → MatchedExisting; no new backlog row.
-    gh.seed_search(
-        "rysweet/Simard",
-        &sig,
-        Ok(vec![GhIssue {
-            number: 42,
-            url: "https://github.com/rysweet/Simard/issues/42".into(),
-            title: "t".into(),
-            body: format!("stewardship-signature: {sig}"),
-        }]),
-    );
-    let second = process_orchestrator_run(&run, &gh, &mut board).unwrap();
-    assert!(matches!(
-        second,
-        StewardshipOutcome::MatchedExisting {
-            issue_number: 42,
-            ..
-        }
-    ));
-    assert_eq!(
-        board.backlog.len(),
-        1,
-        "second call must not duplicate backlog row"
-    );
-}
-
-#[test]
-fn process_run_routes_amplihack_failures_to_amplihack_repo() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let run = amplihack_run();
-    let sig = failure_signature(&run.failure_kind, &run.error_text);
-
-    gh.seed_search("rysweet/amplihack", &sig, Ok(vec![]));
-    gh.seed_create(
-        "rysweet/amplihack",
-        Ok(GhIssue {
-            number: 7,
-            url: "https://github.com/rysweet/amplihack/issues/7".into(),
-            title: "t".into(),
-            body: format!("stewardship-signature: {sig}"),
-        }),
-    );
-
-    let outcome = process_orchestrator_run(&run, &gh, &mut board).unwrap();
-    if let StewardshipOutcome::FiledNew { repo, .. } = outcome {
-        assert_eq!(repo, "rysweet/amplihack");
-    } else {
-        panic!("expected FiledNew");
-    }
-}
-
-// ─────────────────────────── Failure propagation ───────────────────────────
-
-#[test]
-fn process_run_propagates_gh_search_failure() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let run = sample_run();
-    let sig = failure_signature(&run.failure_kind, &run.error_text);
-
-    gh.seed_search(
-        "rysweet/Simard",
-        &sig,
-        Err(SimardError::StewardshipGhCommandFailed {
-            reason: "gh: rate limit exceeded".into(),
-        }),
-    );
-
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipGhCommandFailed { .. }
-    ));
-    assert_eq!(
-        gh.create_call_count(),
-        0,
-        "must not create when search fails"
-    );
-    assert!(board.backlog.is_empty());
-}
-
-#[test]
-fn process_run_propagates_gh_create_failure() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let run = sample_run();
-    let sig = failure_signature(&run.failure_kind, &run.error_text);
-
-    gh.seed_search("rysweet/Simard", &sig, Ok(vec![]));
-    gh.seed_create(
-        "rysweet/Simard",
-        Err(SimardError::StewardshipGhCommandFailed {
-            reason: "gh: 422 validation failed".into(),
-        }),
-    );
-
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipGhCommandFailed { .. }
-    ));
-    assert!(board.backlog.is_empty(), "no backlog row when create fails");
-}
-
-#[test]
-fn process_run_does_not_call_gh_when_routing_ambiguous() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let mut run = sample_run();
-    run.source_module = "totally_unknown_subsystem".to_string();
-
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipRoutingAmbiguous { .. }
-    ));
-    assert_eq!(gh.search_call_count(), 0);
-    assert_eq!(gh.create_call_count(), 0);
-    assert!(board.backlog.is_empty());
-}
-
-// ─────────────────────────── Input validation ───────────────────────────
-
-#[test]
-fn process_run_rejects_empty_run_id() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let mut run = sample_run();
-    run.run_id = String::new();
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipInvalidRunSummary { field } if field == "run_id"
-    ));
-    assert_eq!(gh.search_call_count(), 0);
-}
-
-#[test]
-fn process_run_rejects_empty_source_module() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let mut run = sample_run();
-    run.source_module = String::new();
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipInvalidRunSummary { field } if field == "source_module"
-    ));
-}
-
-#[test]
-fn process_run_rejects_empty_failure_kind() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let mut run = sample_run();
-    run.failure_kind = String::new();
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipInvalidRunSummary { field } if field == "failure_kind"
-    ));
-}
-
-#[test]
-fn process_run_rejects_empty_error_text() {
-    let gh = FakeGhClient::new();
-    let mut board = GoalBoard::new();
-    let mut run = sample_run();
-    run.error_text = String::new();
-    let err = process_orchestrator_run(&run, &gh, &mut board).unwrap_err();
-    assert!(matches!(
-        err,
-        SimardError::StewardshipInvalidRunSummary { field } if field == "error_text"
-    ));
-}
-
-// Suppress unused warnings on the RefCell import (kept for future extensions).
-#[allow(dead_code)]
-fn _unused_refcell_anchor() -> RefCell<()> {
-    RefCell::new(())
-}
